+++ apps/rush-lib/src/data/RushConfiguration.ts
@@ -3,17 +3,15 @@
 
 import * as path from 'path';
 import * as fsx from 'fs-extra';
-import * as os from 'os';
 import * as semver from 'semver';
+import { JsonFile, JsonSchema } from '@microsoft/node-core-library';
 
 import rushVersion from '../rushVersion';
-import JsonFile from '../utilities/JsonFile';
 import RushConfigurationProject, { IRushConfigurationProjectJson } from './RushConfigurationProject';
 import { PinnedVersionsConfiguration } from './PinnedVersionsConfiguration';
 import Utilities from '../utilities/Utilities';
 import { RushConstants } from '../RushConstants';
 import { ApprovedPackagesPolicy } from './ApprovedPackagesPolicy';
-import JsonSchemaValidator from '../utilities/JsonSchemaValidator';
 import EventHooks from './EventHooks';
 import { VersionPolicyConfiguration } from './VersionPolicyConfiguration';
 
@@ -103,6 +101,8 @@ export interface IRushLinkJson {
  */
 export default class RushConfiguration {
   private _rushJsonFile: string;
+  private static _jsonSchema: JsonSchema = JsonSchema.fromFile(path.join(__dirname, '../rush.schema.json'));
+
   private _rushJsonFolder: string;
   private _changesFolder: string;
   private _commonFolder: string;
@@ -146,7 +146,7 @@ export default class RushConfiguration {
    * an RushConfiguration object.
    */
   public static loadFromConfigurationFile(rushJsonFilename: string): RushConfiguration {
-    const rushConfigurationJson: IRushConfigurationJson = JsonFile.loadJsonFile(rushJsonFilename);
+    const rushConfigurationJson: IRushConfigurationJson = JsonFile.load(rushJsonFilename);
 
     // Check the Rush version *before* we validate the schema, since if the version is outdated
     // then the schema may have changed.
@@ -160,15 +160,7 @@ export default class RushConfiguration {
       }
     }
 
-    const rushSchemaFilename: string = path.join(__dirname, '../rush.schema.json');
-    const validator: JsonSchemaValidator = JsonSchemaValidator.loadFromFile(rushSchemaFilename);
-
-    validator.validateObject(rushConfigurationJson, (errorDescription: string) => {
-      const errorMessage: string = `Error parsing file '${rushJsonFilename}':`
-        + os.EOL + errorDescription;
-
-      throw new Error(errorMessage);
-    });
+    RushConfiguration._jsonSchema.validateObject(rushConfigurationJson, rushJsonFilename);
 
     return new RushConfiguration(rushConfigurationJson, rushJsonFilename);
   }