+++ libraries/api-extractor/src/test/DocElementParser.test.ts
@@ -6,10 +6,11 @@
 import { assert } from 'chai';
 import * as ts from 'typescript';
 import * as path from 'path';
+import { JsonFile } from '@microsoft/node-core-library';
+
 import DocElementParser from '../DocElementParser';
 import { IDocElement, IParam, IHrefLinkElement, ICodeLinkElement, ITextElement, ISeeDocElement } from '../IDocElement';
 import TestFileComparer from '../TestFileComparer';
-import JsonFile from '../JsonFile';
 import ApiStructuredType from '../definitions/ApiStructuredType';
 import ApiDocumentation from '../definitions/ApiDocumentation';
 import Extractor from './../Extractor';
@@ -108,8 +109,8 @@ describe('DocElementParser tests', function (): void {
         } as IHrefLinkElement
       ];
       const actualSummary: IDocElement[] = DocElementParser.parse(myDocumentedClass.documentation, tokenizer);
-      JsonFile.saveJsonFile('./lib/basicDocExpected.json', JSON.stringify(expectedSummary));
-      JsonFile.saveJsonFile('./lib/basicDocActual.json', JSON.stringify(actualSummary));
+      JsonFile.save(JSON.stringify(expectedSummary), './lib/basicDocExpected.json');
+      JsonFile.save(JSON.stringify(actualSummary), './lib/basicDocActual.json');
       TestFileComparer.assertFileMatchesExpected('./lib/basicDocActual.json', './lib/basicDocExpected.json');
 
       // Testing Returns Doc Elements
@@ -118,8 +119,8 @@ describe('DocElementParser tests', function (): void {
       ];
       tokenizer.getToken();
       const actualReturn: IDocElement[] = DocElementParser.parse(myDocumentedClass.documentation, tokenizer);
-      JsonFile.saveJsonFile('./lib/returnDocExpected.json', JSON.stringify(expectedReturn));
-      JsonFile.saveJsonFile('./lib/returnDocActual.json', JSON.stringify(actualReturn));
+      JsonFile.save(JSON.stringify(expectedReturn), './lib/returnDocExpected.json');
+      JsonFile.save(JSON.stringify(actualReturn), './lib/returnDocActual.json');
       TestFileComparer.assertFileMatchesExpected('./lib/returnDocActual.json', './lib/returnDocExpected.json');
 
       // Testing Params Doc Elements
@@ -139,8 +140,8 @@ describe('DocElementParser tests', function (): void {
       tokenizer.getToken();
       actualParam.push(apiDoc.parseParam(tokenizer));
 
-      JsonFile.saveJsonFile('./lib/paramDocExpected.json', JSON.stringify(expectedParam));
-      JsonFile.saveJsonFile('./lib/paramDocActual.json', JSON.stringify(actualParam));
+      JsonFile.save(JSON.stringify(expectedParam), './lib/paramDocExpected.json');
+      JsonFile.save(JSON.stringify(actualParam), './lib/paramDocActual.json');
       TestFileComparer.assertFileMatchesExpected('./lib/paramDocActual.json', './lib/paramDocExpected.json');
       assertCapturedErrors([]);
     });
@@ -156,8 +157,8 @@ describe('DocElementParser tests', function (): void {
       ];
       tokenizer.getToken();
       const actualDeprecated: IDocElement[] = DocElementParser.parse(myDocumentedClass.documentation, tokenizer);
-      JsonFile.saveJsonFile('./lib/deprecatedDocExpected.json', JSON.stringify(expectedDeprecated));
-      JsonFile.saveJsonFile('./lib/deprecatedDocActual.json', JSON.stringify(actualDeprecated));
+      JsonFile.save(JSON.stringify(expectedDeprecated), './lib/deprecatedDocExpected.json');
+      JsonFile.save(JSON.stringify(actualDeprecated), './lib/deprecatedDocActual.json');
       TestFileComparer.assertFileMatchesExpected('./lib/deprecatedDocActual.json', './lib/deprecatedDocExpected.json');
       assertCapturedErrors([]);
     });
@@ -184,8 +185,8 @@ describe('DocElementParser tests', function (): void {
           } as ISeeDocElement
       ];
       const actualSummary: IDocElement[] = DocElementParser.parse(myDocumentedClass.documentation, tokenizer);
-      JsonFile.saveJsonFile('./lib/seeDocExpected.json', JSON.stringify(expectedSummary));
-      JsonFile.saveJsonFile('./lib/seeDocActual.json', JSON.stringify(actualSummary));
+      JsonFile.save(JSON.stringify(expectedSummary), './lib/seeDocExpected.json');
+      JsonFile.save(JSON.stringify(actualSummary), './lib/seeDocActual.json');
       TestFileComparer.assertFileMatchesExpected('./lib/seeDocExpected.json', './lib/seeDocActual.json');
       assertCapturedErrors([]);
     });
@@ -215,8 +216,8 @@ describe('DocElementParser tests', function (): void {
       } as IParam;
       const actualParam: IParam = apiDoc.parseParam(tokenizer);
 
-      JsonFile.saveJsonFile('./lib/nestedParamDocExpected.json', JSON.stringify(expectedParam));
-      JsonFile.saveJsonFile('./lib/nestedParamDocActual.json', JSON.stringify(actualParam));
+      JsonFile.save(JSON.stringify(expectedParam), './lib/nestedParamDocExpected.json');
+      JsonFile.save(JSON.stringify(actualParam), './lib/nestedParamDocActual.json');
       TestFileComparer.assertFileMatchesExpected(
         './lib/nestedParamDocActual.json',
         './lib/nestedParamDocExpected.json'