+++ libraries/node-core-library/src/index.ts
@@ -6,4 +6,16 @@
  */
 declare const packageDescription: void; // tslint:disable-line:no-unused-variable
 
-export { JsonFile } from './JsonFile';
+export { FileDiffTest } from './FileDiffTest';
+export {
+  JsonFile,
+  IJsonFileSaveOptions,
+  IJsonFileStringifyOptions
+} from './JsonFile';
+export {
+  JsonSchema,
+  IJsonSchemaErrorInfo,
+  IJsonSchemaValidateOptions,
+  IJsonSchemaFromFileOptions
+} from './JsonSchema';
+export { PackageJsonLookup } from './PackageJsonLookup';