+++ libraries/api-extractor/CHANGELOG.md
@@ -1,6 +1,25 @@
 # Change Log - @microsoft/api-extractor
 
-This log was last generated on Tue, 15 Aug 2017 01:29:31 GMT and should not be manually modified.
+This log was last generated on Thu, 24 Aug 2017 22:44:12 GMT and should not be manually modified.
+
+## 2.3.5
+Thu, 24 Aug 2017 22:44:12 GMT
+
+### Patches
+
+- Update the schema validator.
+
+## 2.3.4
+Thu, 24 Aug 2017 01:04:33 GMT
+
+*Changes not tracked*
+
+## 2.3.3
+Tue, 22 Aug 2017 13:04:22 GMT
+
+### Patches
+
+- Added "api-documenter" code sample
 
 ## 2.3.2
 Tue, 15 Aug 2017 01:29:31 GMT